Added a temporary one-page Apple Help Book, which mostly contains just a link to macsvg.org for now.
Updated README.md to show a new path animation example, the macSVG logo animation file.

Added a function to convert nearly everything in a path to cubic bezier segments.  The exceptions are the path commands M, Z and A (move, close path, and elliptical arc).  In some situations, this is helpful for designing path animation.

The animation timeline is now sorted by the computed value of the "begin" attribute.  For "animate" elements, the "attributeName" value is shown in the timeline.

Removed my user name from the remote file server login xib files.