Several improvements were made for editing <image> elements, with changes in MacSVGDocument and ImageElementEditor.
A new attribute was added: "xlink:role".  It contains the URL for the image.  In many situations, it will duplicate the value of the URL in the "xlink:href" attribute.  But if the "xlink:href" attribute is used as a container for embedding the image data, the original URL will be retained in the "xlink:role" attribute.

A new button was added to the Image Element Editor - "Set Image xlink:href Attribute".  It will set the attribute value according to the radio buttons above to the selected setting: "Link to Image", "Embed PNG", or "Embed JPEG".

It is now possible to replace an image in a selected image element by dragging the new image into the image well, and clicking the "Set Image xlink:href Attribute" button.  The image element width and height will be adjusted to match the new image.