The functions provided in IRIReferenceAttributeEditor plugin are now merged into the SVGAttributeEditor plugin, and the IRIReferenceEditor is removed from the macSVG application bundle.
The IRI Reference attributes (e.g., "url(#radialGradient1)") are now appended to the "Defined Values" popup button menu, with a separator item between the predefined values, and the IRI references found in the document.

The prototypes for the <animate>, <animateColor>, <animateMotion>, and <animateTransform> elements changed the default value for the "fill" attribute from "freeze" to "remove".