This commit has several changes for drawing completion of polylines, polygons and paths.  These elements can be completed by double-clicking the mouse, or pressing the escape, return or space key on the keyboard.
In the path Element Editor, the "Close Path Automatically" will automatically add a "Z" or "z" command to the path data to close the path.  If the current path mode is set to "Cubic Curve" with absolute coordinates, a final cubic curve ("C" command) will automatically added to the beginning of the path, preceding the "Z" command.  A bug still exists if a different item is selected in the outline view while drawing a polyline or polygon - the last point is incorrectly deleted - so the other methods for completing those shapes are recommended until that bug is fixed.

In the web view, the "dblClick" event is no longer registered as an event to be listened for.  Instead, the mouseUp compares the event timestamp with the previous mouseUp timestamp, and handles the double-click operations.

In XMLOutlineController, the outlineViewItemWillExpand: added a @synchronized directive as a mutex lock to block handleOutlineViewSelectionDidChange.