The drawing tools for <path> elements were improved to simplify the creation of subpaths for effects like the "donut-hole" in a closed and filled path shape.
The easiest way to create a subpath while drawing a path element is to press the option key while using one of the various methods to close a path (for example, clicking the "Close Path" button).  The current subpath with be closed, and a new subpath will be started, ready for the next click in the web view to set the starting point of the new subpath.

Add a "fill-rule" attribute to the path element to choose a rule: "nonzero" or "evenodd".  The nonzero rule removes subpaths according to the "winding" of the path - clockwise or counterclockwise.  The "evenodd" rule is useful for some paths with overlapping segments.