Added ConvertPathsDialog.plugin for the Plug-Ins menu.  Initially, this plugin contains one option to convert all selected SVG <path> elements to absolute coordinates, which is sometimes useful with imported SVG clipart to improve the usability of the path editing controls.  More path conversion options will likely be added in the future.
Notable, this is the first macSVG Cocoa plugin to use Swift language rather than Objective-C.  The Swift libraries are now added to the main app bundle.  Some plugin import statements and header interfaces were modified for Swift compatibility.  Swift seems to work okay for plugins, but more changes may be required to expose all of the plugin callbacks to the host application in Swift.

Additional work is underway to migrate to Swift, including testing of a Swift-first application plugin architecture, available separately at this repository: https://github.com/dsward2/PluginHost

The current plan is to rewrite most of the application-specific code with Swift, and replace the third-party network code - the built-in HTTP server and SFTP upload/download features - with libraries acquired with the Swift Package Manager.