In MacSVGDocuments, added a new method named countAllXMLElements.
When opening a new document, or reverting to an existing document, the document elements are counted.  If 100 or fewer elements are contained in the document, the XMLOutlineController expandAllNodes method is called.  When a SVG document with more that 100 elements is loaded, the outline view is collapsed to the top element.