These changes make it easier to edit path, line, polyline and polygon elements that are translated, rotated and scaled with a "transform" attribute.
In point-editing mode with the crosshair cursor, the handles for the end points and control points in the web view will now use the element's transform values, including transforms inherited from parent elements (typically the <g> group element).

For example, if a path element is rotated, the path editing handles should now appear at the correct locations on the rotated path, and mouse edits in the web view are easier to perform.