This commit has initial changes to allow element and attribute plug-in editors to display function-specific editing handles in the web view.
The TransformAttributeEdtor plugin uses this feature to display a center-of-rotation handle when editing a "rotate" command in a "transform" attribute.

The next commit is planned to allow the new type of handle to be interactive like the other existing handle types.  For editing the rotate command in a transform attribute, this will allow the mouse to drag the center of rotation to a new position.