This commit has several changes for transform operations.
The new center-of-rotation handle for rotate transform is now interactive in the web view.  You can use the mouse to click-and-drag the center-of-rotation handle, with live updates displayed for the rotated objects.

Some partial support is added for drawing tools to create objects into a selected group with the group's current transform, initially the path, polyline, polygon and line drawing tools.  The rect, circle, ellipse, text and image elements remain to be done in a future commit.  For most objects, editing an existing objects should use transforms better now.