This revision changes the ruler views alongside the SVG web view.  A new RulerView class was defined, and it is now the superclass for VerticalRulerView and HorizontalRulerView.  Each RulerView now contains a separate web view for drawing the ruler markers with SVG markup, and the major increments of the ruler are labeled with a <text> element (0, 100, 200, etc.).
Currently, the rulers are drawn and labeled in "px" units, but provisions are made for other units like "cm" and "in" in the future.  Since the rulers are essentially web views, that should make it easier to match the ruler with the main web view's computation of unit space.

Currently, the rulers do not translate and scale properly when the main web view is zoomed or scrolled away from the origin.  The previous rulers did not work properly either.  The ruler translate and scale will be fixed in a future commit.