This commit brings improvements to drag-and-drop functions in the XML outline view, and a change in the user interface usage.
Drag-and-drop operations for multiple items should work better now, without requiring the earlier workaround of pressing the control key during dragging.  The improvements also work for drag-and-drops for items from one document window to another one.

The user interface change is: when multiple items are selected in the outline view, and the user wishes to select a single item within the group of currently selected items, double-click on that single item to select it.  Subsequent selections can be made with a single click.  The up and down arrow keys on the keyboard remain available to navigate the outline view selection to the previous or next item.