These changes are intended to improve the initial settings for the AnimateTransform element.  When a new AnimateTransform element is added to the SVG document, the parent element of the AnimateTransform element will be checked for a "transform" attribute.  If the transform attribute is present in the parent element, the last transform operation will be used to set the attributes in the AnimateTransform element accordingly.
For example, if the parent element's transform attribute is set to "rotate(45, 37, 58)", the "type" attribute in the AnimateTransform element will be set to "rotate", and the "values" attribute in the AnimateTransform will be based on the parent element's rotate values.  For the animation, the starting degrees value will match the parent's rotate value, and the center of rotation will match the parent's.  An additional set of values will be added to rotate the parent element a full 360 degrees.

The other transform operations (translate, scale, skewX, skewY) will produce similar results for new AnimateTransform elements.  For a "matrix" transform, the starting and ending matrix values are the same, but can be modified in the editor after the AnimateTransform element is added.