This revision allows the center-of-rotation marker used in "transform" attributes to be used in sync with an "animateTransform" element.
When the center-of-rotation marker is dragged, this plug-in will search the selected element child elements for an animateTransform element.  If an animateTransform element is found with type="rotate", the center-of-rotation coordinates are compared between the selected element and the animateTransform element.  If the coordinates match, the animateTransform center-of-rotation coordinates will be updated to match the coordinates for the selected element.