In SVGPathEditor modifyPath method, the mouse controls for setting the radius of a new elliptical arc segment are improved.
To use this function, when clicking the mouse to set the end point for a new elliptical arc segment, hold and drag the mouse to change the radius value.  The effect of changing the radius will depend on the "Large Arc" checkbox setting.

The SVG specification allows independent values for x and y radius, but in this implementation, those are set to an equal value, so the shape of the arc will be circular when using this dragging command to set the arc radius.