In the AnimateTransform plugin, a stepper control was added for the "dur" attribute value text field.  The AnimateTransform plugin preference score was adjusted for a few attributes - type, repeatCount, calcMode.
A bug was fixed concerning plugin editors that are valid for both element and attribute editing (like AnimateTransform).  The bug caused some plugin panels to fail to load when selected with the menu.

A bug was fixed in the Element Info plugin, which was causing the text data to fail to display.