The MARKETING_VERSION is incremented to v 1.1.6.
All cell-based table views were converted to view-based.

A standard macOS Share button was added to the tool panel.  It is mainly useful for sending an AirDrop URL to an iPhone or Mac to quickly access the built-in macSVG web server.  It produces some warning messages in the log file about SHKRemoteView, but they seem to be common for AirDrop sharing applications and are apparently harmless.